pyproject.toml
[tool.poetry]
name = "chainlit-langgraph"
version = "0.1.0"
description = ""
authors = ["Bruce Chou <brucechou1983@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.1"
langgraph = "^0.2.31"
langchain-openai = "^0.2.1"
langchain-anthropic = "^0.2.1"
langchain-ollama = "^0.2.0"
langchain-google-vertexai = "^2.0.1"
chainlit = "1.3.1"
boto3 = "^1.35.41"
alembic = "^1.13.3"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
langchain-community = "^0.3.2"
pypdf = "^5.0.1"
langchain-groq = "^0.2.1"
langchain-google-genai = "^2.0.4"
google-auth-oauthlib = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
commitizen = "^3.30.0"
filelock = "^3.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.3.1"
tag_format = "v$version"
changelog_incremental = true
update_changelog_on_bump = true

[tool.commitizen.version_files]
"chat_workflow/version.py" = [
    '__version__ = "{version}"'
]
"README.md" = [
    'Version-{version}-blue'
]
